thread: add custom stack size
An option to request different stack size for a seastar thread
is added - via thread_attributes struct. Note that the actual
stack size depends on the configuration (e.g. the presence of ASAN
or arch-dependent config), so what users pass via thread_attributes
is only a *requested* custom stack size.
A use case in mind is allowing lighter threads to be used for operations
which do not expect large stacks, but are expected to be on the fast path.
A concrete example is handling alternator requests in Scylla, where
threads could help with reducing reactor stalls induced by rapidjson
parsing.