rpc: add log levels
In order to be able to log with different log levels,
another flavor of set_logger() function is added - based on
seastar's log levels from utils/log.hh.
In order to preserve backward compatibility, the legacy level-less
logger is still usable. In order to use the new logger, simply
use the new overload of the set_logger function:

    _rpc->set_logger([] (log_level level, std::string_view str) {
            rpc_logger.log(level, "{}", log);
    });

The compatibility rules are as follows:
1. If the user still defines a legacy level-less logger,
   Seastar will internally assume that it should print messages
   with severity greater than or equal to INFO. All internal usages
   were checked to assure that the behavior would not change.
2, If the user only defines the new logger, but still uses
   logging without explicit level, INFO will be assumed.